+++ core/services/ocr2/plugins/ccip/internal/ccipdata/factory/offramp_test.go
@@ -15,19 +15,19 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils"
 	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0"
+	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0"
 )
 
 func TestOffRamp(t *testing.T) {
-	for _, versionStr := range []string{ccipdata.V1_0_0, ccipdata.V1_2_0} {
+	for _, versionStr := range []string{ccipdata.V1_2_0} {
 		lggr := logger.Test(t)
 		addr := cciptypes.Address(utils.RandomAddress().String())
 		lp := mocks2.NewLogPoller(t)
 
 		expFilterNames := []string{
-			logpoller.FilterName(v1_0_0.EXEC_EXECUTION_STATE_CHANGES, addr),
-			logpoller.FilterName(v1_0_0.EXEC_TOKEN_POOL_ADDED, addr),
-			logpoller.FilterName(v1_0_0.EXEC_TOKEN_POOL_REMOVED, addr),
+			logpoller.FilterName(v1_2_0.EXEC_EXECUTION_STATE_CHANGES, addr),
+			logpoller.FilterName(v1_2_0.EXEC_TOKEN_POOL_ADDED, addr),
+			logpoller.FilterName(v1_2_0.EXEC_TOKEN_POOL_REMOVED, addr),
 		}
 		versionFinder := newMockVersionFinder(ccipconfig.EVM2EVMOffRamp, *semver.MustParse(versionStr), nil)
