+++ core/services/ocr2/plugins/ccip/internal/ccipdata/offramp_reader_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/chains/evm/utils"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/commit_store_helper"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_offramp"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_offramp_1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_offramp_1_2_0"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/mock_rmn_contract"
 	"github.com/smartcontractkit/chainlink/v2/core/internal/testutils"
@@ -32,54 +31,12 @@ import (
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipcalc"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/factory"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0"
 )
 
 type offRampReaderTH struct {
 	user   *bind.TransactOpts
 	reader ccipdata.OffRampReader
-}
-
-func TestExecOnchainConfig100(t *testing.T) {
-	tests := []struct {
-		name      string
-		want      v1_0_0.ExecOnchainConfig
-		expectErr bool
-	}{
-		{
-			name: "encodes and decodes config with all fields set",
-			want: v1_0_0.ExecOnchainConfig{
-				PermissionLessExecutionThresholdSeconds: rand.Uint32(),
-				Router:                                  utils.RandomAddress(),
-				PriceRegistry:                           utils.RandomAddress(),
-				MaxTokensLength:                         uint16(rand.Uint32()),
-				MaxDataSize:                             rand.Uint32(),
-			},
-		},
-		{
-			name: "encodes and fails decoding config with missing fields",
-			want: v1_0_0.ExecOnchainConfig{
-				PermissionLessExecutionThresholdSeconds: rand.Uint32(),
-				MaxDataSize:                             rand.Uint32(),
-			},
-			expectErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			encoded, err := abihelpers.EncodeAbiStruct(tt.want)
-			require.NoError(t, err)
-
-			decoded, err := abihelpers.DecodeAbiStruct[v1_0_0.ExecOnchainConfig](encoded)
-			if tt.expectErr {
-				require.ErrorContains(t, err, "must set")
-			} else {
-				require.NoError(t, err)
-				require.Equal(t, tt.want, decoded)
-			}
-		})
-	}
 }
 
 func TestExecOnchainConfig120(t *testing.T) {
@@ -130,14 +87,6 @@ func TestOffRampReaderInit(t *testing.T) {
 		version string
 	}{
 		{
-			name:    "OffRampReader_V1_0_0",
-			version: ccipdata.V1_0_0,
-		},
-		{
-			name:    "OffRampReader_V1_1_0",
-			version: ccipdata.V1_1_0,
-		},
-		{
 			name:    "OffRampReader_V1_2_0",
 			version: ccipdata.V1_2_0,
 		},
@@ -181,11 +130,6 @@ func setupOffRampReaderTH(t *testing.T, version string) offRampReaderTH {
 	// Setup offRamp.
 	var offRampAddress common.Address
 	switch version {
-	case ccipdata.V1_0_0:
-		offRampAddress = setupOffRampV1_0_0(t, user, bc)
-	case ccipdata.V1_1_0:
-		// Version 1.1.0 uses the same contracts as 1.0.0.
-		offRampAddress = setupOffRampV1_0_0(t, user, bc)
 	case ccipdata.V1_2_0:
 		offRampAddress = setupOffRampV1_2_0(t, user, bc)
 	case ccipdata.V1_5_0:
@@ -205,42 +149,6 @@ func setupOffRampReaderTH(t *testing.T, version string) offRampReaderTH {
 		user:   user,
 		reader: reader,
 	}
-}
-
-func setupOffRampV1_0_0(t *testing.T, user *bind.TransactOpts, bc *client.SimulatedBackendClient) common.Address {
-	onRampAddr := utils.RandomAddress()
-	armAddr := deployMockArm(t, user, bc)
-	csAddr := deployCommitStore(t, user, bc, onRampAddr, armAddr)
-
-	// Deploy the OffRamp.
-	staticConfig := evm_2_evm_offramp_1_0_0.EVM2EVMOffRampStaticConfig{
-		CommitStore:         csAddr,
-		ChainSelector:       testutils.SimulatedChainID.Uint64(),
-		SourceChainSelector: testutils.SimulatedChainID.Uint64(),
-		OnRamp:              onRampAddr,
-		PrevOffRamp:         common.Address{},
-		ArmProxy:            armAddr,
-	}
-	sourceTokens := []common.Address{}
-	pools := []common.Address{}
-	rateLimiterConfig := evm_2_evm_offramp_1_0_0.RateLimiterConfig{
-		IsEnabled: false,
-		Capacity:  big.NewInt(0),
-		Rate:      big.NewInt(0),
-	}
-
-	offRampAddr, tx, offRamp, err := evm_2_evm_offramp_1_0_0.DeployEVM2EVMOffRamp(user, bc, staticConfig, sourceTokens, pools, rateLimiterConfig)
-	bc.Commit()
-	require.NoError(t, err)
-	ccipdata.AssertNonRevert(t, tx, bc, user)
-
-	// Verify the deployed OffRamp.
-	tav, err := offRamp.TypeAndVersion(&bind.CallOpts{
-		Context: testutils.Context(t),
-	})
-	require.NoError(t, err)
-	require.Equal(t, "EVM2EVMOffRamp 1.0.0", tav)
-	return offRampAddr
 }
 
 func setupOffRampV1_2_0(t *testing.T, user *bind.TransactOpts, bc *client.SimulatedBackendClient) common.Address {
